+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Changed
+
+- Move third party services to adapters [(#90)](https://github.com/plotwist-app/plotwist-backend/pull/90)
+
+### Added
 - Add review summary [(#87)](https://github.com/plotwist-app/plotwist-backend/pull/87)
 
 ## 1.9.0