+++ CHANGELOG.md
@@ -14,8 +14,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Add service to run in docker compose file too [(#88)](https://github.com/plotwist-app/plotwist-backend/pull/89)
-- Consume SQS messages [(#71)](https://github.com/plotwist-app/plotwist-backend/pull/71)
 - Add review summary [(#87)](https://github.com/plotwist-app/plotwist-backend/pull/87)
+- Consume SQS messages [(#71)](https://github.com/plotwist-app/plotwist-backend/pull/71)
 
 ## 1.9.0