+++ CHANGELOG.md
@@ -5,14 +5,17 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+## 1.10.0
+
 ### Changed
 
 - Move third party services to adapters [(#90)](https://github.com/plotwist-app/plotwist-backend/pull/90)
 
 ### Added
 
-- Consume SQS messages [(#71)](https://github.com/plotwist-app/plotwist-backend/pull/71)
+- Add service to run in docker compose file too [(#88)](https://github.com/plotwist-app/plotwist-backend/pull/89)
 - Add review summary [(#87)](https://github.com/plotwist-app/plotwist-backend/pull/87)
+- Consume SQS messages [(#71)](https://github.com/plotwist-app/plotwist-backend/pull/71)
 
 ## 1.9.0